refactor: 프롬프트 관련 파일을 prompt 하위 패키지로 이동
question 패키지 내부에 혼재되어 있던 프롬프트 관련 파일들을
prompt 하위 패키지로 분리하여 도메인 구조를 명확히 정리함.

- PromptController, PromptService, Prompt 엔티티, DTO, 테스트 등 이동
- question 패키지는 질문 도메인 전용으로 유지

Related to: modajigi/interview_mate#14
